FIxes and directory expansion
