+++ Enums/ParcelStatus.cs
@@ -1,0 +1,8 @@
+namespace Inpost_org.Enums;
+
+public enum ParcelStatus
+{
+    InWarehouse,
+    InTransport,
+    Delivered
+}
