+++ Enums/Permission.cs
@@ -1,0 +1,10 @@
+namespace Inpost_org.Enums;
+
+public enum Permission
+{
+    DeletePackages,
+    OrderPackages,
+    DeliverPackages,
+    UpdatePackages,
+    ManageUsers
+}
