+++ Enums/Role.cs
@@ -1,0 +1,8 @@
+namespace Inpost_org.Enums;
+
+public enum Role
+{
+    Administrator,
+    InpostEmployee,
+    InpostClient
+}
