+++ Services/Operations/EventListener.cs
@@ -8,6 +8,11 @@ public class LogManager()
 {
     private static readonly string logFilePath = "event_log.txt";
     
+    /// <summary>
+    /// This method writes message from event to event_log.txt file with current date, time and
+    /// status of the operation.
+    /// </summary>
+    /// <param name="e">MongoDBOperationEventArgs</param>
     protected static void LogToFile(MongoDBOperationEventArgs e)
     {
         string action = e.Success ? "success" : "failure";
@@ -20,6 +25,14 @@ public class LogManager()
 
 public class EventListener : LogManager
 {
+    /// <summary>
+    /// Handles user-related operations and logs the result.
+    /// Displays a success or failure message in the console with appropriate formatting.
+    /// Logs the operation details to a file.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="person">The user involved in the operation.</param>
+    /// <param name="e">Event arguments containing operation details and status.</param>
     public static void OnUserOperation(object sender, PersonModel person, MongoDBOperationEventArgs e)
     {
         string user = person == null ? "Admin" : person.Username;
@@ -46,6 +59,15 @@ public class EventListener : LogManager
         LogToFile(e);
     }
 
+    /// <summary>
+    /// Handles parcel-related operations and logs the result.
+    /// Displays a success or failure message in the console with appropriate formatting.
+    /// Logs the operation details to a file.
+    /// </summary>
+    /// <param name="sender">The source of the event.</param>
+    /// <param name="parcel">The parcel involved in the operation.</param>
+    /// <param name="person">The user associated with the parcel operation.</param>
+    /// <param name="e">Event arguments containing operation details and status.</param>
     public static void OnParcelOperation(object sender, ParcelModel parcel, PersonModel person, MongoDBOperationEventArgs e)
     {
         string action = e.Success ? "success" : "failure";
