+++ Services/Operations/ParcelOperations/ShowParcelOperation.cs
@@ -1,11 +1,22 @@
+using Inpost_org.Enums;
 using Inpost_org.Services.NotificationMethods;
 using Inpost_org.Users.Deliveries;
 using Inpost_org.Users;
 
 namespace Inpost_org.Services.Operations.ParcelOperations;
 
+/// <summary>
+/// Class responsible for displaying details of a specific parcel from the MongoDB database.
+/// </summary>
 public class ShowParcelOperation : ParcelBase
 {
+    /// <summary>
+    /// Executes the operation of displaying details of a specific parcel.
+    /// </summary>
+    /// <param name="mongo">The MongoDB service object.</param>
+    /// <param name="parcel">The parcel model representing the parcel to be displayed.</param>
+    /// <param name="person">The person model representing the user performing the operation.</param>
+    /// <param name="e">The event arguments for the MongoDB operation.</param>
     public override void Operation(MongoDBService mongo, ParcelModel parcel, PersonModel person, MongoDBOperationEventArgs e)
     {
         e.Operation = "ShowParcel";
