+++ Services/Operations/UserOperations/ShowParcelsOperation.cs
@@ -1,3 +1,4 @@
+using Inpost_org.Enums;
 using Inpost_org.Services.NotificationMethods;
 using Inpost_org.Users;
 using Inpost_org.Users.Deliveries;
@@ -6,8 +7,18 @@ using MongoDB.Driver;
 
 namespace Inpost_org.Services.Operations.UserOperations;
 
+/// <summary>
+/// Class responsible for displaying parcels associated with a user or accessible by their role.
+/// </summary>
 public class ShowParcelsOperation : UserBase
 {
+    /// <summary>
+    /// Executes the operation of displaying parcels associated with the specified user or accessible by their role.
+    /// </summary>
+    /// <param name="mongo">The MongoDB service object.</param>
+    /// <param name="person">The person model representing the user requesting the operation.</param>
+    /// <param name="e">The event arguments for the MongoDB operation.</param>
+    /// <param name="role">The role of the user requesting the operation in case of Admin.</param>
     public override void Operation(MongoDBService mongo, PersonModel person, MongoDBOperationEventArgs e, string role)
     {
         e.Operation = "Show parcels";
