+++ Services/Operations/UserOperations/UpdateUserOperation.cs
@@ -4,8 +4,18 @@ using MongoDB.Driver;
 
 namespace Inpost_org.Services.Operations.UserOperations;
 
+/// <summary>
+/// Class responsible for updating multiple users' details in the MongoDB database.
+/// </summary>
 public class UpdateUserOperation : UserBase
 {
+    /// <summary>
+    /// Executes the operation of updating multiple users' details in the database.
+    /// </summary>
+    /// <param name="mongo">The MongoDB service object.</param>
+    /// <param name="person">The person model representing the user requesting the operation.</param>
+    /// <param name="e">The event arguments for the MongoDB operation.</param>
+    /// <param name="role">The role used to filter or validate the users to be updated.</param>
     public override void Operation(MongoDBService mongo, PersonModel person, MongoDBOperationEventArgs e, string role)
     {
         e.Operation = "UpdateUser";
