+++ Users/Deliveries/ParcelModel.cs
@@ -1,17 +1,17 @@
+using Inpost_org.Enums;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
 
 namespace Inpost_org.Users.Deliveries;
 
-public enum ParcelStatus
-{
-    InWarehouse,
-    InTransport,
-    Delivered
-}
-
+/// <summary>
+/// Represents a parcel in the system, including its name, sender, recipient, and status.
+/// </summary>
 public class ParcelModel
 {
+    /// <summary>
+    /// The unique identifier for the parcel, stored as an ObjectId in MongoDB.
+    /// </summary>
     [BsonId]
     [BsonRepresentation(BsonType.ObjectId)]
     public string Id { get; set; }
@@ -20,6 +20,11 @@ public class ParcelModel
     public PersonModel Recipient { get; set; }
     public ParcelStatus Status { get; set; }
     
+    /// <summary>
+    /// Initializes a new instance of the <see cref="ParcelModel"/> class with the specified name and recipient.
+    /// </summary>
+    /// <param name="parcelName">The name of the parcel.</param>
+    /// <param name="recipient">The recipient of the parcel.</param>
     public ParcelModel(string parcelName, PersonModel recipient)
     {
         ParcelName = parcelName;
